oh man. you would think at some point i would learn to not do dumb things like use a smallint on a database for a zip code....

+++ doretax/biz/models.py
@@ -88,7 +88,7 @@ class BusinessDetail(models.Model):
     address = models.CharField(max_length=100)
     city = models.CharField(max_length=100)
     state = USStateField(choices=STATE_CHOICES)
-    postal_code = models.SmallIntegerField(blank=True, null=True, help_text="(optional)")
+    postal_code = models.IntegerField(blank=True, null=True, help_text="(optional)")
     
     def __unicode__(self):
         return self.name
